refactor: 분리된 Controller 통합
각 Card TableView 마다 Controller를 가지고 있던 것을
하나의 Controller 안에 3개의 TableView를 가지도록 수정
