--- ToDoListApp/ToDoListApp/InProgressController.swift
@@ -1,29 +1,0 @@
-//
-//  InProgressViewController.swift
-//  ToDoListApp
-//
-//  Created by user on 2021/04/07.
-//
-
-import UIKit
-
-class InProgressController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
